The resource manager initialization works! (I mean it does nothing but still)

+++ kernel/resource/resource_manager.h
@@ -28,6 +28,6 @@ error_t resource_manager_get_provider_by_resource(struct process* process, resou
 
 error_t resource_manager_get_provider_by_scheme(const char* scheme, resource_provider_t** provider);
 
-error_t resource_manager_get_provider_by_pid(int pid, resource_provider_t** provider);
+error_t resource_manager_get_provider_by_pid(size_t pid, resource_provider_t** provider);
 
 #endif
